+++ remix/app/routes/peers.tsx
@@ -1,0 +1,5 @@
+import { DeploymentPeersPage } from '~/components/Peers/DeploymentPeersPage';
+
+export default function DeploymentPeersRoute() {
+	return <DeploymentPeersPage />;
+}